Updated code accordingly to Iulian's code review.
