--- org.scala-ide.sdt.core.tests/src/scala/tools/eclipse/package.scala
@@ -1,10 +1,0 @@
-package scala.tools
-
-import org.eclipse.jdt.core.ICompilationUnit
-import scala.tools.eclipse.javaelements.ScalaCompilationUnit
-
-package object eclipse {
-  implicit def iCompUnit2ScalaCompUnit(unit: ICompilationUnit): ScalaCompilationUnit = {
-    unit.asInstanceOf[ScalaCompilationUnit]
-  }
-}
