+++ org.scala-ide.sdt.core.tests/src/scala/tools/eclipse/testsetup/TestProjectSetup.scala
@@ -5,8 +5,17 @@ import org.eclipse.jdt.core.IPackageFragmentRoot
 import org.eclipse.jdt.core.JavaCore
 import org.junit.Assert.assertNotNull
 import scala.tools.eclipse.ScalaProject
+import org.eclipse.jdt.core.ICompilationUnit
+import scala.tools.eclipse.javaelements.ScalaSourceFile
 import scala.tools.eclipse.javaelements.ScalaCompilationUnit
 import org.eclipse.jdt.core.ICompilationUnit
+import org.eclipse.jdt.core.IProblemRequestor
+import org.eclipse.jdt.core.WorkingCopyOwner
+import scala.tools.eclipse.javaelements.ScalaCompilationUnit
+import scala.tools.eclipse.javaelements.ScalaSourceFile
+import org.eclipse.core.runtime.NullProgressMonitor
+
+import org.mockito.Mockito.{mock, when}
 
 /** Base class for setting up tests that depend on a project found in the test-workspace.
  * 
@@ -44,8 +53,8 @@ class TestProjectSetup(projectName: String) {
     srcPackageRoot.getPackageFragment(segments.init.mkString(".")).getCompilationUnit(segments.last)
   }
   
-  def scalaCompilationUnit(path: String): ScalaCompilationUnit =
-    compilationUnit(path).asInstanceOf[ScalaCompilationUnit]
+  def scalaCompilationUnit(path: String): ScalaCompilationUnit with ICompilationUnit =
+    compilationUnit(path).asInstanceOf[ScalaSourceFile]
   
   def reload(unit: ScalaCompilationUnit) {
     // first, 'open' the file by telling the compiler to load it
@@ -62,5 +71,43 @@ class TestProjectSetup(projectName: String) {
     	val contents = unit.getContents()
     	SDTTestUtils.positionsOf(contents, marker)
     }
+  }
+  
+  /** Emulate the opening of a scala source file (i.e., it tries to 
+   * reproduce the steps performed by JDT when opening a file in an editor). 
+   * 
+   * @param srcPath the path to the scala source file 
+   * */
+  def open(srcPath: String) = {
+    val unit = scalaCompilationUnit(srcPath)
+    openWorkingCopyFor(unit)
+    reload(unit)
+    unit
+  }
+  
+  /** Open a working copy of the passed `unit` */
+  private def openWorkingCopyFor(unit: ScalaCompilationUnit with ICompilationUnit) {
+    val requestor = mock(classOf[IProblemRequestor])
+    // the requestor must be active, or unit.getWorkingCopy won't trigger the Scala
+    // structure builder
+    when(requestor.isActive()).thenReturn(true)
+
+    val owner = new WorkingCopyOwner() {
+      override def getProblemRequestor(unit: org.eclipse.jdt.core.ICompilationUnit): IProblemRequestor = requestor
+    }
+
+    // this will trigger the Scala structure builder
+    unit.getWorkingCopy(owner, new NullProgressMonitor)
+  }
+  
+  /** Wait until the passed `unit` is entirely typechecked. */
+  def waitUntilTypechecked(unit: ScalaCompilationUnit) {
+    // give a chance to the background compiler to report the error
+    project.withSourceFile(unit) { (source, compiler) =>
+      import scala.tools.nsc.interactive.Response
+      val res = new Response[compiler.Tree]
+      compiler.askLoadedTyped(source, res)
+      res.get // wait until unit is typechecked
+    }()
   }
 }
